# docs/swagger.yaml
basePath: /v1
definitions:
  service.UserLoginService:
    properties:
      password:
        type: string
      user_name:
        type: string
    required:
    - password
    - user_name
    type: object
  service.UserRegisterService:
    properties:
      nickname:
        type: string
      password:
        type: string
      password_confirm:
        type: string
      user_name:
        type: string
    required:
    - nickname
    - password
    - password_confirm
    - user_name
    type: object
host: api.kyuuu.be
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample cms server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: QB blog Swagger API
  version: "1.0"
paths:
  /api/v1/user/list:
    get:
      consumes:
      - application/json
      description: get user list
      parameters:
      - description: 关键字
        in: query
        name: keyWords
        type: string
      - description: 用户角色
        in: query
        name: role
        type: string
      - description: 注册时间
        in: query
        name: creatTime
        type: string
      produces:
      - application/json
      summary: 获取用户列表
      tags:
      - accounts
  /api/v1/user/login:
    post:
      consumes:
      - application/json
      description: get accounts
      parameters:
      - description: login account
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.UserLoginService'
          type: object
      produces:
      - application/json
      summary: 用户登录接口
      tags:
      - accounts
  /api/v1/user/me:
    get:
      consumes:
      - application/json
      description: get current uer info
      parameters:
      - description: 用户id
        in: query
        name: id
        type: string
      produces:
      - application/json
      summary: 获取用户信息
      tags:
      - accounts
  /api/v1/user/register:
    post:
      consumes:
      - application/json
      description: register account
      parameters:
      - description: register account
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.UserRegisterService'
          type: object
      produces:
      - application/json
      summary: 用户注册接口
      tags:
      - accounts
swagger: "2.0"